Refactor: Move hydra task model adapter to model module
